refactor(storefront): split the default storefront into named views
Storefront.tsx held the hero, the category rail, the search box and the
catalogue rows in one 371-line file. Each is now its own component under
components/storefront/default/, and Storefront.tsx is the wiring that
joins the controller to the default view.

DefaultStorefront is a pure view: everything it renders arrives as props.
That makes it the reference implementation of the storefront contract, so
a bespoke outlet component can be read next to it.

HeroBento no longer returns null below three featured games. A new outlet
with two curated games rendered an empty band with no explanation; the
main tile now widens to fill the row when there are no side tiles.

Adds the data-storefront hooks (search, filters, game-list, game-link)
that the contract guard will assert on.

Rendering is byte-identical: screenshots of /store, /store/[slug], and its
category and search states match the previous commit exactly.
