+++ components/store/GameListItem.tsx
@@ -25,6 +25,7 @@ export function GameListItem({
   return (
     <Link
       href={itemHref}
+      data-storefront="game-link"
       className="group block rounded-3xl border border-white/10 bg-white/5 p-0 md:p-4 shadow-sm backdrop-blur transition-all duration-300 hover:shadow-[0_0_30px_rgba(165,180,252,0.1)] hover:border-white/20 motion-safe:hover:-translate-y-1 relative overflow-hidden"
     >
       <div className="flex items-stretch gap-2 md:gap-6">
